+++ genesis-creator/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased changes
 
-- Update `rust-sdk` dependency so that project is forward compatible.
+- Update `concordium-rust-sdk` dependency so that project is forward compatible.
 
 ## 0.5.0